% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_accuracy}
\alias{data_accuracy}
\title{Accuracy data}
\format{A data frame with 48000 obs of 5 variables.}
\usage{
data(data_accuracy)
}
\description{
A dataset of comparisons between software in terms of
accuracy of diagrams constructed from sampling set intersections
}
\details{
\describe{
\item{it}{an integer variable of the number of the iteration}
\item{Sets}{A factor of the number of sets used in the diagram}
\item{software}{A factor of the software package}
\item{Metric}{A factor of the type of loss (diagError or stress)}
\item{Loss}{The loss of the solution}
}
}
\author{
Johan Larsson
}
\keyword{datasets}
