% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_accuracy_int}
\alias{data_accuracy_int}
\title{Accuracy data for three-set intersections}
\format{A data frame with 14000 obs of 4 variables.}
\usage{
data(data_accuracy_int)
}
\description{
A dataset of comparisons between software packages in terms of
accuracy in reproducing three-sets layouts with all intersections present
}
\details{
\describe{
\item{it}{an integer variable of the number of the iteration}
\item{software}{A factor of the software package}
\item{Metric}{A factor of the type of loss (diagError or stress)}
\item{Loss}{The loss of the solution}
}
}
\author{
Johan Larsson
}
\keyword{datasets}
