% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_consistency}
\alias{data_consistency}
\title{Consistency data}
\format{A data frame with 1800 obs of 4 variables.}
\usage{
data(data_consistency)
}
\description{
A dataset of comparisons between venneuler and eulerr in terms of
reproducibility of diagrams constructed from sampling circles,
and thus diagrams that have perfect solutions.
}
\details{
#' \describe{
\item{One}{First item}
\item{Two}{Second item}
}
\itemize{
\item Sets. A factor of the number of sets used in the diagram
\item it. An integer variable of the number of the iteration
\item software. A factor with two levels, signifying which software (eulerr or
venneuler) that was used.
\item stress. A double of the stress metric from venneuler.
}
}
\author{
Johan Larsson
}
\keyword{datasets}
