% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_performance}
\alias{data_performance}
\title{Performance data}
\format{A data frame with 18720 obs of 4 variables.}
\usage{
data(data_accuracy)
}
\description{
A dataset of comparisons between software packages in terms of
computational performance in generating
diagrams constructed from sampling set intersections.
}
\details{
\describe{
\item{it}{an integer variable of the number of the iteration}
\item{software}{The software (method) used. A factor variable.}
\item{Sets}{The number of sets in the diagram}
\item{Time}{The computing speed in nanoseconds.}
}
}
\author{
Johan Larsson
}
\keyword{datasets}
