% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{degenerate_split_conics}
\alias{degenerate_split_conics}
\title{Form a degenerate conic from two ellipses and then split it into two lines}
\usage{
degenerate_split_conics(C1, C2)
}
\arguments{
\item{C1}{Ellipse 1}

\item{C2}{Ellipse 2}
}
\value{
Two lines
}
\description{
Form a degenerate conic from two ellipses and then split it into two lines
}
