% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{diag_error}
\alias{diag_error}
\title{diagError}
\usage{
diag_error(orig, fit)
}
\arguments{
\item{orig}{Original values}

\item{fit}{Fitted values}
}
\value{
diagError
}
\description{
diagError
}
