% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{ellipse_arc}
\alias{ellipse_arc}
\title{Ellipse arcs}
\usage{
ellipse_arc(saxes = c(1, 1), loc = c(0, 0), theta = 0, n = 200,
  rng = c(0, 2 * pi))
}
\arguments{
\item{saxes}{Semi-axes}

\item{loc}{Center}

\item{theta}{Rotation}

\item{n}{Number of vertices}

\item{rng}{Range of the arc to return.}
}
\value{
Points on the arc of an ellipse.
}
\description{
Ellipse arcs
}
