% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr2017bsc.R
\docType{package}
\name{eulerr2017bsc-package}
\alias{eulerr2017bsc}
\alias{eulerr2017bsc-package}
\title{eulerr2017bsc: Source Code for a 2017 Bachelor's Thesis on the R-Package eulerr}
\description{
Source Code for a 2017 bachelor's thesis in statistics
about the R-package eulerr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/eulerr2017bsc}
  \item Report bugs at \url{https://github.com/jolars/eulerr2017bsc/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com}

Authors:
\itemize{
  \item Peter Gustafsson \email{peter.gustafsson@stat.lu} (supervisor)
}

}
\keyword{internal}
