% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_impossible}
\alias{fig_impossible}
\title{A set relationship visualized with circles and ellipses}
\usage{
fig_impossible()
}
\value{
Plots two plots side-by-side.
}
\description{
A set relationship visualized with circles and ellipses
}
