% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_intersection}
\alias{fig_intersection}
\title{The process of intersecting two ellipses}
\usage{
fig_intersection()
}
\value{
Plot three diagrams side-by-side showing the process of intersecting
  three ellipses.
}
\description{
The process of intersecting two ellipses
}
