% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_polyarea}
\alias{fig_polyarea}
\title{Example of overlap computation for ellipses}
\usage{
fig_polyarea()
}
\value{
Plot a showcase of overlap area computations.
}
\description{
Example of overlap computation for ellipses
}
