% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_skyline}
\alias{fig_skyline}
\title{Shocase the SKYLINE-BL aglorithm for packing rectangles (ellipses)}
\usage{
fig_skyline()
}
\value{
A series of plots showcasing how the skyline algorithm works.
}
\description{
Shocase the SKYLINE-BL aglorithm for packing rectangles (ellipses)
}
