% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_venneulerHard}
\alias{fig_venneulerHard}
\title{Plots a comparison of the hard set in venneuler paper}
\usage{
fig_venneulerHard()
}
\value{
A side-by-side comparison.
}
\description{
Plots a comparison of the hard set in venneuler paper
}
