% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{gof}
\alias{gof}
\alias{gof.VennDrawing}
\alias{gof.VennDiagram}
\title{Goodness of fit tests for euler diagram}
\usage{
gof(obj, orig)

\method{gof}{VennDrawing}(obj, orig)

\method{gof}{VennDiagram}(obj, orig)
}
\arguments{
\item{obj}{A fit euler diagram}
}
\description{
Goodness of fit tests for euler diagram
}
