% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{plot.VennDiagram}
\alias{plot.VennDiagram}
\title{Hijacks the venneuler plot method}
\usage{
\method{plot}{VennDiagram}(x, col, col.fn = function(col) hcl(col * 360, 130,
  60), alpha = 0.3, main = NULL, edges = 200, border = NA,
  col.txt = 1, ...)
}
\arguments{
\item{x}{An object of class 'VennDiagram'}

\item{col}{Color postions on the gradient}

\item{col.fn}{Color function to produce colors with}

\item{alpha}{Opacity}

\item{main}{Titles}

\item{edges}{Ignored}

\item{border}{Border color}

\item{col.txt}{ignored}

\item{...}{Passed on to [plot.euler]}
}
\value{
Plots a venneuler diagram using eulerr methods.
}
\description{
Hijacks the venneuler plot method to plot using grid instead of
base graphics, provide appropriate limits and label positions.
}
