% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{plot_vennjs}
\alias{plot_vennjs}
\title{Plot venn.js fit}
\usage{
plot_vennjs(obj, fill = c("#1f77b4", "#ff7f0e", "#2ca02c", "#d62728",
  "#9467bd", "#8c564b", "#e377c2", "#7f7f7f", "#bcbd22", "#17becf"),
  col = fill, alpha = 0.3, edges = 200, border = NA, col.txt = 1, ...)
}
\arguments{
\item{obj}{An object of class 'vennjs'}

\item{col}{color}

\item{alpha}{Opacity}

\item{edges}{Ignored}

\item{border}{Border color}

\item{col.txt}{ignored}

\item{...}{Passed on to \link{plot.euler}}

\item{labels}{a character string of labels}
}
\value{
Plot a diagram from venn.js
}
\description{
Plot venn.js fit
}
