% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_test}
\alias{prop_test}
\title{Hypothesis test of two-sample proportions}
\usage{
prop_test(p0, p1, n0, n1)
}
\arguments{
\item{p0}{proportion of sample 1}

\item{p1}{proportion of sample 2}

\item{n0}{sample size of sample 1}

\item{n1}{sample size of sample 2}
}
\value{
a p-value
}
\description{
Hypothesis test of two-sample proportions
}
