% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_vennjs}
\alias{setup_vennjs}
\title{Setup input for vennjs fitting}
\usage{
setup_vennjs(combinations)
}
\arguments{
\item{combinations}{Disjoint set relationships}
}
\value{
A character vector of set relationships that venn.js can work with.
}
\description{
Setup input for vennjs fitting
}
