% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_consistency}
\alias{data_consistency}
\title{Consistency data}
\format{A data frame with 226336 obs of 6 variables.}
\usage{
data(data_consistency)
}
\description{
A dataset of comparisons between Euler diagram methods in the
reproducibility of diagrams constructed from sampling circles,
and thus diagrams that have perfect solutions.
}
\details{
\describe{
\item{it}{The iteration. An integer.}
\item{shape}{The shape of the diagram. A factor.}
\item{sets}{The number of sets. A factor (integer) variable.}
\item{software}{The software used in the observation. A factor.}
\item{metric}{The loss metric. A factor.}
\item{loss}{The loss of the function. A double vector.}
}
}
\author{
Johan Larsson
}
\keyword{datasets}
